remove un-needed deps. add run script for launching CexControl
